Moved xrange series into subfolder.
