+++ js/masters/modules/xrange.src.js
@@ -10,4 +10,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Series/XRangeSeries.js';
+import '../../Series/XRange/XRangeSeries.js';
